Introduce Subscription Filters
This PR extends the Subscription API to allow invoking side effects, similar to the Query API. With this enhancement,
the following functionalities are now supported:

Filter execution upon network reconnection:

* `SubscriptionOptions#restartOnReconnect`
* `SwrCachePlusPolicy#networkResumeSubscriptionsFilter`

Filter execution upon window refocus:

* `SubscriptionOptions#restartOnFocus`
* `SwrCachePlusPolicy#windowResumeSubscriptionsFilter`

Triggering side effects via SwrClient:

* `SwrClient#effect`

Triggering side effects upon successful mutation:

* `MutationKey#onMutateEffect`

+++ soil-query-core/src/commonMain/kotlin/soil/query/SubscriptionCommand.kt
@@ -36,7 +36,7 @@ interface SubscriptionCommand<T> {
 }
 
 internal typealias SubscriptionErrorRelay = (ErrorRecord) -> Unit
-internal typealias SubscriptionRestart = () -> Unit
+internal typealias SubscriptionRestart = (Long) -> Unit
 
 /**
  * Dispatches the result of the subscription.
